feat: doctor cli command prints duplciate uuids
I recently broke parts of my org setup by refactoring some folders
together and combining things with roam - roam now does me the service
of throwing a huge error when trying to process some of my files,
because there are a large number of duplicated uuids in there. That part
is my fault, and it's an annoying problem to have, especially when it
means a big buffer printing a huge group of uuids and saying i thinks
that at least one is a duplicate.

This commit introduces a cli command 'doctor', whos first task is
analyzing your org-directory for problems, specifically duplicate uuids.
Duplicates are logged. If this is common or there are quite alot, maybe
we'll add some interactive support for updating the particular items,
'Carve' style.