refactor: remove dynamic vars
I never liked this api - this removes all dynamic vars. Some of these
were optional/flexible in a way that was
unnecessary (item->source-file), and others are better as a default. If
an org property is repeated in the same node, it will come out as a
list. It used to bug me that some items would have a scalar, others a
list, but it's dynamic and easy enough to deal with, plus it's quite
rare - i don't expect to use it. It came from a time that i was using
org as a database and needed more flexible metadata, but that's just
crazy, we're dropping that use-case.