feat: restore create/refile and tests
Refactors the :add-item portion of the update flow to both add to a
matching org item's children _and_ append to the aggregated :items. This
is necessary after skip-children in item->lines. There's some
ambiguity/confusion between the flattened and nested item approaches
that could probably be simpler - for now we continue to support both.