feat: pull in organum rewrite, drop organum dep
The organum dep org-crud was built on was rewritten to use instaparse -
i was hoping to do the same and drop the dep. This pulls in the rewrite
in a separate namespace. Next is dropping/melding these implementations
together and reducing/cleaning up/dedupling the work in
org-crud/node.clj.